+++ doc/ideatools_advanced_fr.R
@@ -4,6 +4,15 @@ knitr::opts_chunk$set(
   comment = "#>"
 )
 
+## ----echo = FALSE-------------------------------------------------------------
+scales::show_col(c(
+  "tres defavorable" = "#CD0000",
+  "defavorable" = "#FF6347",
+  "intermédiaire" = "#FCD400",
+  "favorable" = "#33FF00",
+  "tres favorable" = "#008B00"
+), borders = TRUE)
+
 ## ----eval = FALSE-------------------------------------------------------------
 #  install.packages("remotes")
 #  remotes::install_github("davidcarayon/IDEATools")
@@ -22,12 +31,12 @@ scales::show_col(c("Agroécologique"= "#2e9c15", "Socio-Territoriale" = "#5077FE
 
 ## ----echo = FALSE-------------------------------------------------------------
 scales::show_col(c(
-  "tres defavorable" = "#CD0000",
-  "defavorable" = "#FF6347",
-  "intermédiaire" = "#FCD400",
-  "favorable" = "#33FF00",
-  "tres favorable" = "#008B00"
-), borders = TRUE)
+  "tres defavorable" = "#FF0000",
+  "defavorable" = "#FF6348",
+  "intermédiaire" = "#FFA300",
+  "favorable" = "#1CDA53",
+  "tres favorable" = "#0D8A00"
+), borders = FALSE)
 
 ## ----out.width = "100%", echo = FALSE-----------------------------------------
 knitr::include_graphics("img/pipeline_individuel.png")